Move task provider into own folder
